+++ src/Evaluator/EvalError.php
@@ -12,10 +12,13 @@ use Serhii\GoodbyeHtml\Obj\ObjType;
 
 readonly class EvalError
 {
+    private const PREFIX = '[EVAL_ERROR]';
+
     public static function wrongArgumentType(string $type, ObjType $expect, Obj $actual): ErrorObj
     {
         return new ErrorObj(sprintf(
-            '[EVAL_ERROR] "%s" is not allowed argument type for "%s", expected "%s"',
+            '%s "%s" is not allowed argument type for "%s", expected "%s"',
+            self::PREFIX,
             $actual->type()->value,
             $type,
             $expect->value,
@@ -24,14 +27,15 @@ readonly class EvalError
 
     public static function unknownType(Node $node): ErrorObj
     {
-        return new ErrorObj('[EVAL_ERROR] unknown type: "' . get_class($node) . '"');
+        return new ErrorObj(sprintf('%s unknown type: "%s"', self::PREFIX, get_class($node)));
     }
 
     public static function operatorNotAllowed(string $operator, Obj $right): ErrorObj
     {
         return new ErrorObj(
             sprintf(
-                '[EVAL_ERROR] operator "%s" is not allowed for type "%s"',
+                '%s operator "%s" is not allowed for type "%s"',
+                self::PREFIX,
                 $operator,
                 $right->type()->value,
             )
@@ -40,14 +44,15 @@ readonly class EvalError
 
     public static function variableIsUndefined(VariableExpression $node): ErrorObj
     {
-        return new ErrorObj(sprintf('[EVAL_ERROR] variable "$%s" is undefined', $node->value));
+        return new ErrorObj(sprintf('%s variable "$%s" is undefined', self::PREFIX, $node->value));
     }
 
     public static function infixExpressionMustBeBetweenNumbers(string $side, string $operator, Obj $obj): ErrorObj
     {
         return new ErrorObj(
             sprintf(
-                '[EVAL_ERROR] %s side must be INT or FLOAT for operator %s. Got %s',
+                '%s %s side must be INT or FLOAT for operator %s. Got %s',
+                self::PREFIX,
                 $side,
                 $operator,
                 $obj->type()->value,